hoshino/modules/majsoul/requirements.txt
bs4
urllib3
pillow
